refactor: 유틸클래스와 메서드명 이름을 좀 더 명확하게 수정한다
